+++ src/special/generalized-harmonic.js
@@ -12,7 +12,7 @@ import hurwitzZeta from './hurwitz-zeta'
  * @private
  */
 export default function (n, m) {
-  if (n < 10) {
+  if (n < 20) {
     // If n is small, just calculate it exactly.
     let z = 0
     for (let k = 1; k <= n; k++) {